+++ src/main/java/com/merlin/digitalbanking/ebankingbackend/entities/SavingAccount.java
@@ -6,11 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @DiscriminatorValue("SA")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class SavingAccount extends BankAccount{
-    private double interestRate;
+    private BigDecimal interestRate;
 }
